Consolidate errors and types into a new entry point
