+++ .gitignore
@@ -3,3 +3,4 @@
 /lib
 /node_modules
 package-lock.json
+types.js
