+++ src/core/swap/faast-plugin.js
@@ -3,12 +3,6 @@
 import { gt, lt } from 'biggystring'
 
 import {
-  SwapAboveLimitError,
-  SwapBelowLimitError,
-  SwapCurrencyError,
-  SwapPermissionError
-} from '../../types/error.js'
-import {
   type EdgeCurrencyWallet,
   type EdgePluginEnvironment,
   type EdgeSpendInfo,
@@ -16,7 +10,11 @@ import {
   type EdgeSwapPlugin,
   type EdgeSwapPluginQuote,
   type EdgeSwapQuoteOptions,
-  type EdgeSwapTools
+  type EdgeSwapTools,
+  SwapAboveLimitError,
+  SwapBelowLimitError,
+  SwapCurrencyError,
+  SwapPermissionError
 } from '../../types/types.js'
 import { makeSwapPluginQuote } from './swap-helpers.js'
