+++ src/index.js
@@ -16,22 +16,4 @@ export { destroyAllContexts } from './core/root.js'
 export { makeBrowserIo } from './io/browser/browser-io.js'
 export { makeNodeIo } from './io/node/node-io.js'
 export { makeReactNativeIo } from './io/react-native/react-native-dummy.js'
-export {
-  DustSpendError,
-  errorNames,
-  InsufficientFundsError,
-  SpendToSelfError,
-  NetworkError,
-  NoAmountSpecifiedError,
-  ObsoleteApiError,
-  OtpError,
-  PasswordError,
-  PendingFundsError,
-  SameCurrencyError,
-  SwapAboveLimitError,
-  SwapBelowLimitError,
-  SwapCurrencyError,
-  SwapPermissionError,
-  UsernameError
-} from './types/error.js'
 export { makeEdgeContext, makeFakeContexts } from './makeContext.js'
